Jump and duck are changed, shadows are also objects now
Lights are removed because of the low performance, shadows are also
objects under the other objects.

+++ HyperSpace/Assets/Prefabs/Materials/Transparent Shader.shader
@@ -44,9 +44,9 @@
                 return o;
             }
            
-            fixed4 frag (output i) : COLOR
+            half4 frag (output i) : COLOR
             {
-                fixed4 col = tex2D(_MainTex, i.uv);
+                half4 col = tex2D(_MainTex, i.uv);
  
                 col *= _Color;
